feat: update

+++ scripts/publish-beta-extension.ts
@@ -19,24 +19,26 @@ const ossClient = oss({
 function updateBetaDependencies(extension: string, directory: string) {
   try {
     const publishedPackages: string[] = JSON.parse(fs.readFileSync(path.join(__dirname, 'publishedPackages.temp.json'), 'utf-8'));
-    const packageFile = path.join(directory, 'package.json');
-    const packageData = fs.readJsonSync(packageFile);
 
-    publishedPackages.forEach((publishedPackage: string) => {
-      const info = publishedPackage.split(':');
-      const name = info[0];
-      const version = info[1];
+    if (fs.existsSync(directory)) {
+      const packageFile = path.join(directory, 'package.json');
+      const packageData = fs.readJsonSync(packageFile);
 
-      if (packageData.dependencies[name]) {
-        packageData.dependencies[name] = version;
-      } else if (packageData.devDependencies[name]) {
-        packageData.devDependencies[name] = version;
-      }
-    });
-    fs.writeFileSync(packageFile, JSON.stringify(packageData, null, 2));
+      publishedPackages.forEach((publishedPackage: string) => {
+        const info = publishedPackage.split(':');
+        const name = info[0];
+        const version = info[1];
 
+        if (packageData.dependencies[name]) {
+          packageData.dependencies[name] = version;
+        } else if (packageData.devDependencies[name]) {
+          packageData.devDependencies[name] = version;
+        }
+      });
+      fs.writeFileSync(packageFile, JSON.stringify(packageData, null, 2));
+    }
   } catch (e) {
-    console.log(`[ERROR] ${extension} upload beta package dependencies failed.`, e);
+    console.log(`[ERROR] ${extension} update beta package dependencies failed.`, e);
   }
 };
 
@@ -77,7 +79,11 @@ getExtensionInfos().then((extensionInfos: IExtensionInfo[]) => {
     if (shouldPublish) {
       publishedCount++;
       console.log(`--- ${name}@${localVersion} ---`);
+      
       updateBetaDependencies(name, directory);
+      // Update inside web project package json
+      updateBetaDependencies(name,  path.join(directory, 'web'));
+
       publish(name, directory, localVersion);
       publishedExtensions.push(`${name}:${localVersion}`);
     }
